requirements.txt
fastapi==0.128.0
passlib==1.7.4
psycopg2_binary==2.9.11
pydantic==2.12.5
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
